+++ go/border/rpkt/route.go
@@ -23,14 +23,12 @@ import (
 
 	"github.com/scionproto/scion/go/border/metrics"
 	"github.com/scionproto/scion/go/border/rcmn"
-	"github.com/scionproto/scion/go/border/rctx"
 	"github.com/scionproto/scion/go/lib/addr"
 	"github.com/scionproto/scion/go/lib/assert"
 	"github.com/scionproto/scion/go/lib/common"
 	"github.com/scionproto/scion/go/lib/overlay"
 	"github.com/scionproto/scion/go/lib/ringbuf"
 	"github.com/scionproto/scion/go/lib/scmp"
-	"github.com/scionproto/scion/go/lib/topology"
 	"github.com/scionproto/scion/go/proto"
 )
 
@@ -266,31 +264,4 @@ func (rp *RtrPkt) reprocess() (HookResult, error) {
 	s.Ring.Write(ringbuf.EntryList{rp}, true)
 	// Stop routing the packet after enqueuing it back into the ringbuffer.
 	return HookFinish, nil
-}
-
-// getSVCNamesMap returns the slice of instance names and addresses for a given SVC address.
-func getSVCNamesMap(svc addr.HostSVC, ctx *rctx.Ctx) ([]string,
-	map[string]topology.TopoAddr, error) {
-
-	t := ctx.Conf.Topo
-	var names []string
-	var elemMap map[string]topology.TopoAddr
-	switch svc.Base() {
-	case addr.SvcBS:
-		names, elemMap = t.BSNames, t.BS
-	case addr.SvcPS:
-		names, elemMap = t.PSNames, t.PS
-	case addr.SvcCS:
-		names, elemMap = t.CSNames, t.CS
-	case addr.SvcSB:
-		names, elemMap = t.SBNames, t.SB
-	default:
-		return nil, nil, common.NewBasicError("Unsupported SVC address",
-			scmp.NewError(scmp.C_Routing, scmp.T_R_BadHost, nil, nil), "svc", svc)
-	}
-	if len(elemMap) == 0 {
-		return nil, nil, common.NewBasicError("No instances found for SVC address",
-			scmp.NewError(scmp.C_Routing, scmp.T_R_UnreachHost, nil, nil), "svc", svc)
-	}
-	return names, elemMap, nil
 }
